Fix SettingsManagerImpl
